+++ src/models/requestLoggerFactoryOptions.ts
@@ -1,0 +1,11 @@
+export interface RequestLoggerFactoryOptions {
+  useShort?: boolean;
+  requestOutput?: boolean;
+  requestHeaders?: boolean;
+  requestBodyLength?: number;
+  responseHeaders?: boolean;
+  responseBodyPrettyPrint?: boolean;
+  responseBodyLength?: number;
+  timings?: boolean;
+  onlyFailed?: boolean;
+}